Fixed bug with clearing mcID if loggin from Guest to Logged-in mode

+++ CordialSDK/CordialSDK/Endpoints/UpsertContacts/Helper/ContactsSender.swift
@@ -14,10 +14,12 @@ class ContactsSender {
     let upsertContacts = UpsertContacts()
     
     func upsertContacts(upsertContactRequests: [UpsertContactRequest]) {
-        
-        if let previousPrimaryKey = CordialUserDefaults.string(forKey: API.USER_DEFAULTS_KEY_FOR_PREVIOUS_PRIMARY_KEY) {
+                
+        if InternalCordialAPI().isUserLogin() {
+            let previousPrimaryKey = CordialUserDefaults.string(forKey: API.USER_DEFAULTS_KEY_FOR_PREVIOUS_PRIMARY_KEY)
+            
             upsertContactRequests.forEach { upsertContactRequest in
-                if upsertContactRequest.primaryKey != previousPrimaryKey {
+                if upsertContactRequest.primaryKey != previousPrimaryKey && previousPrimaryKey != nil {
                     CoreDataManager.shared.deleteAllCoreData()
                     InternalCordialAPI().removeCurrentMcID()
                 }
